requirements-mac.txt
flask
flask-cors
# pyautogui (Removed for Headless Safety)
# pywhatkit (Removed for Headless)
# pyperclip (Removed for Headless)
psutil
google-generativeai
python-dotenv
duckduckgo-search
opencv-python-headless
numpy<2.0.0
pyinstaller
pillow
pydantic
google-genai
pydantic
groq
requests
